JSON BITCHES

+++ MGChat/Components/MovableComponent.cs
@@ -2,8 +2,8 @@
 {
     public class MovableComponent : ECS.Component
     {
-        public int Speed;
-        public MovableComponent(int parent, int speed=1) : base(parent)
+        public float Speed;
+        public MovableComponent(int parent, float speed=1) : base(parent)
         {
             Speed = speed;
         }
